requirements.txt
fastapi
uvicorn[standard]
pydub
google-genai
python-json-logger
pre-commit
ruff
black
python-multipart
audioop-lts
google-cloud-pubsub
python-dotenv
google-auth


google-cloud-storage
psycopg
google-cloud-aiplatform
numpy
pandas
